refactor: Name the two IDE-shared packages for what they are
vscode-bridge becomes webview-bridge (@dineug/erd-editor-webview-bridge) and
vscode-replication-store-worker becomes replication-store-worker
(@dineug/erd-editor-replication-store-worker). Both serve the IntelliJ webview as much as
the VSCode one, so the vscode prefix said less than the repo's own vocabulary does: every
IDE-embedded UI here is a webview, and every off-thread module is a worker named for what
it runs. Directories, package names, imports, task-input globs, the inline-worker host
pattern, the workspace file and the guidance move together; the worker name strings and
the vscode-webview, vscode-extension and vuerd-vscode packages keep their names.
