+++ README.md
@@ -137,8 +137,8 @@ highlighting). Everything else is internal.
 | [`app`](./packages/app) | The React PWA at erd-editor.io |
 | [`vscode-extension`](./packages/vscode-extension) | The published VS Code extension |
 | [`vscode-webview`](./packages/vscode-webview) | The bundle inside the VS Code webview |
-| [`vscode-bridge`](./packages/vscode-bridge) | Typed host ↔ webview command protocol |
-| [`vscode-replication-store-worker`](./packages/vscode-replication-store-worker) | Headless document replica for the VS Code host |
+| [`webview-bridge`](./packages/webview-bridge) | Typed host ↔ webview command protocol |
+| [`replication-store-worker`](./packages/replication-store-worker) | Headless document replica for the VS Code host |
 | [`intellij-webview`](./packages/intellij-webview) | The bundle inside the IntelliJ plugin's editor panel |
 | [`intellij-plugin`](./packages/intellij-plugin) | The published IntelliJ plugin — Kotlin and Gradle, not TypeScript |
