+++ packages/intellij-plugin/AGENTS.md
@@ -41,7 +41,7 @@ The Kotlin/JVM half of the JetBrains plugin: a thin host layer that registers a 
 
 - **Publishing is manual.** No token or signing key lives in this repository: build the zip and upload it. Bump `pluginVersion` and fill the matching `CHANGELOG.md` section first.
 - **Dependencies come from the platform** — Jackson, `kotlinx.coroutines` and `org.cef.*` are all bundled, and adding a library here invites classloader conflicts. The build uses Kotlin 2.3.21, but `apiVersion = KOTLIN_2_1` limits compiled code to the Kotlin 2.1 API surface bundled by the 2025.2 floor, and `kotlin.stdlib.default.dependency = false` prevents shipping another stdlib copy. That floor is also why it resolves `intellijIdea(...)` and never `intellijIdeaCommunity(...)`: the separate IC distribution ended after 2025.2.
-- **Bridge command `type` strings must match `packages/vscode-bridge` exactly.** Change one side only and messages are dropped in silence.
+- **Bridge command `type` strings must match `packages/webview-bridge` exactly.** Change one side only and messages are dropped in silence.
 - **Threading:** file writes go inside `readAndEdtWriteAction { writeAction { … } }` (the older `readAndWriteAction` is deprecated), file dialogs open on the EDT through `invokeLater`.
 - `vp staged` globs `**/*.{ts,mts,tsx}`, so a Kotlin-only commit passes pre-commit unchecked. The root `.gitignore` also carries a bare `build` pattern — a Kotlin package named `build` would be untracked while compiling fine locally.
 
@@ -66,7 +66,7 @@ The Kotlin/JVM half of the JetBrains plugin: a thin host layer that registers a 
 
 ### Internal
 
-`@dineug/erd-editor-intellij-webview` produces `src/main/resources/assets`; the `@dineug/erd-editor-vscode-bridge` command definitions are the contract `WebviewBridge.kt` mirrors.
+`@dineug/erd-editor-intellij-webview` produces `src/main/resources/assets`; the `@dineug/erd-editor-webview-bridge` command definitions are the contract `WebviewBridge.kt` mirrors.
 
 ### External
