+++ packages/intellij-webview/AGENTS.md
@@ -6,14 +6,14 @@
 ## Purpose
 
 The HTML/JS bundle embedded by the IntelliJ ERD Editor plugin, whose Kotlin/JVM side is `packages/intellij-plugin`.
-Same `<erd-editor>` element and same `@dineug/erd-editor-vscode-bridge` protocol as `vscode-webview`, but webview→host is
+Same `<erd-editor>` element and same `@dineug/erd-editor-webview-bridge` protocol as `vscode-webview`, but webview→host is
 `window.cefQuery` with a JSON string, so payloads must survive `JSON.stringify`/`parse` and binary data is base64-encoded first.
 
 ## Key Files
 
 | File | Description |
 | --- | --- |
-| `src/main.ts` | The whole client — element creation, `bridge` (host) + `workerBridge` (worker), `createReplicationStoreWorker` from `@dineug/erd-editor-vscode-replication-store-worker`, theme/readonly commands, lazy shiki registration |
+| `src/main.ts` | The whole client — element creation, `bridge` (host) + `workerBridge` (worker), `createReplicationStoreWorker` from `@dineug/erd-editor-replication-store-worker`, theme/readonly commands, lazy shiki registration |
 | `src/env.d.ts` | Ambient `window.cefQuery` / `cefQueryCancel` types and `declare module '*.css'` |
 | `index.html` | Build entry at the package root; `publicDir: false`, so no static asset dir sits beside it |
 | `vite.config.ts` | The one output path, the `stripCrossorigin` plugin, the `worker` block, `run.tasks` |
@@ -33,7 +33,7 @@ Same `<erd-editor>` element and same `@dineug/erd-editor-vscode-bridge` protocol
 - **`base` stays `/`, only `.html`/`.js`/`.css` may be emitted, `sourcemap: false`** — the plugin's CEF
   `SchemeHandlerFactory` maps the URL path onto the classpath and types those three extensions, nothing else.
 - **Do not restore `crossorigin` on injected tags.** `stripCrossorigin` removes it: the scheme handler sends no CORS headers, so the module script is refused and the panel stays blank.
-- **Every worker loads from its URL, and none is written here.** The plugin serves `index.html` and every asset from one origin, `https://<DOMAIN>`, through its scheme handler, so the `new Worker(new URL('./workers/…', import.meta.url))` that `@dineug/erd-editor`, `@dineug/erd-editor-shiki-worker` and `@dineug/erd-editor-vscode-replication-store-worker` ship resolves same-origin and Vite emits each as `static/js/<name>.<hash>.js` — a `.js` the handler's whitelist admits. Nothing is inlined, unlike `vscode-webview`. The replica worker that used to live under `src/services/` is that shared package now.
+- **Every worker loads from its URL, and none is written here.** The plugin serves `index.html` and every asset from one origin, `https://<DOMAIN>`, through its scheme handler, so the `new Worker(new URL('./workers/…', import.meta.url))` that `@dineug/erd-editor`, `@dineug/erd-editor-shiki-worker` and `@dineug/erd-editor-replication-store-worker` ship resolves same-origin and Vite emits each as `static/js/<name>.<hash>.js` — a `.js` the handler's whitelist admits. Nothing is inlined, unlike `vscode-webview`. The replica worker that used to live under `src/services/` is that shared package now.
 - **The `worker` block repeats the output naming** because workers do not inherit `build.rolldownOptions.output`, and `worker.format: 'es'`
   matches `main.ts` building its worker as `{ type: 'module' }`. No `build.target` is set — JCEF's Chromium is the only browser here.
 
@@ -59,7 +59,7 @@ cd ../intellij-plugin && ./gradlew runIde
 
 ### Internal
 
-`@dineug/erd-editor` (element), `@dineug/erd-editor-vscode-bridge` (protocol, reused despite the name), `@dineug/erd-editor-vscode-replication-store-worker` (the replica worker, reused despite the name), `@dineug/erd-editor-shiki-worker` (lazy).
+`@dineug/erd-editor` (element), `@dineug/erd-editor-webview-bridge` (protocol), `@dineug/erd-editor-replication-store-worker` (the replica worker), `@dineug/erd-editor-shiki-worker` (lazy).
 
 ### External
