+++ packages/intellij-webview/package.json
@@ -12,8 +12,8 @@
   "dependencies": {
     "@dineug/erd-editor": "workspace:*",
     "@dineug/erd-editor-shiki-worker": "workspace:*",
-    "@dineug/erd-editor-vscode-bridge": "workspace:*",
-    "@dineug/erd-editor-vscode-replication-store-worker": "workspace:*",
+    "@dineug/erd-editor-webview-bridge": "workspace:*",
+    "@dineug/erd-editor-replication-store-worker": "workspace:*",
     "base64-arraybuffer": "^1.0.2"
   },
   "devDependencies": {
