+++ packages/replication-store-worker/README.md
@@ -1,13 +1,13 @@
-# vscode-replication-store-worker
+# replication-store-worker
 
 > Headless replica of the editor document, running off the main thread
 
-`@dineug/erd-editor-vscode-replication-store-worker` keeps a second copy of the open document inside
+`@dineug/erd-editor-replication-store-worker` keeps a second copy of the open document inside
 a Web Worker, so the VS Code host can persist `.erd.json` files without serializing on the UI thread.
 The webview stays free to render; the worker does the replaying and the stringifying.
 
 Internal to the erd-editor monorepo and never published to npm — depend on it as
-`"@dineug/erd-editor-vscode-replication-store-worker": "workspace:*"`.
+`"@dineug/erd-editor-replication-store-worker": "workspace:*"`.
 `@dineug/erd-editor-vscode-webview` is the only consumer; `intellij-webview` keeps its own copy of
 the worker, spawned from a URL instead of inlined.
 
@@ -17,7 +17,7 @@ The webview forwards the editor's raw action stream into the worker, which repla
 built by `createReplicationStore` from `@dineug/erd-editor/engine.js`. Only when that store reports a
 `change` does the worker post the serialized document back out.
 
-Traffic in both directions is plain `postMessage` carrying `@dineug/erd-editor-vscode-bridge`
+Traffic in both directions is plain `postMessage` carrying `@dineug/erd-editor-webview-bridge`
 commands — inbound `webviewInitialValueCommand` and `webviewReplicationCommand`, outbound
 `hostSaveValueCommand`.
 
@@ -29,8 +29,8 @@ webpack and Rspack emit as a worker file of its own; a host that cannot load a w
 origins, such as the VSCode webview, inlines it in its own build instead.
 
 ```ts
-import { Bridge, hostSaveValueCommand } from '@dineug/erd-editor-vscode-bridge';
-import { createReplicationStoreWorker } from '@dineug/erd-editor-vscode-replication-store-worker';
+import { Bridge, hostSaveValueCommand } from '@dineug/erd-editor-webview-bridge';
+import { createReplicationStoreWorker } from '@dineug/erd-editor-replication-store-worker';
 
 const worker = createReplicationStoreWorker({ name: 'replication-store-worker' });
 const bridge = new Bridge();
@@ -51,5 +51,5 @@ is no separate worker file to ship, and no module resolution at runtime.
 No test task; verify with a build.
 
 ```sh
-pnpm exec vp run --filter @dineug/erd-editor-vscode-replication-store-worker --fail-if-no-match build
+pnpm exec vp run --filter @dineug/erd-editor-replication-store-worker --fail-if-no-match build
 ```
