+++ packages/replication-store-worker/src/services/replicationStore.worker.ts
@@ -5,7 +5,7 @@ import {
   hostSaveValueCommand,
   webviewInitialValueCommand,
   webviewReplicationCommand,
-} from '@dineug/erd-editor-vscode-bridge';
+} from '@dineug/erd-editor-webview-bridge';
 
 import { toWidth } from '@/utils/text';
