+++ packages/replication-store-worker/src/utils/text.ts
