+++ packages/replication-store-worker/src/vite-env.d.ts
