+++ packages/replication-store-worker/vite.config.ts
