+++ packages/vscode-extension/package.json
@@ -48,7 +48,7 @@
     "e2e:typecheck": "tsc -p tsconfig.integration.json --noEmit"
   },
   "devDependencies": {
-    "@dineug/erd-editor-vscode-bridge": "workspace:*",
+    "@dineug/erd-editor-webview-bridge": "workspace:*",
     "@dineug/erd-editor-vscode-webview": "workspace:*",
     "@types/mocha": "^10.0.10",
     "@types/node": "^20.19.0",
