+++ packages/vscode-webview/src/index.ts
@@ -6,6 +6,7 @@ import {
   setGetShikiServiceCallback,
   setImportFileCallback,
 } from '@dineug/erd-editor';
+import { createReplicationStoreWorker } from '@dineug/erd-editor-replication-store-worker';
 import {
   AnyAction,
   Appearance,
@@ -22,8 +23,7 @@ import {
   webviewReplicationCommand,
   webviewUpdateReadonlyCommand,
   webviewUpdateThemeCommand,
-} from '@dineug/erd-editor-vscode-bridge';
-import { createReplicationStoreWorker } from '@dineug/erd-editor-vscode-replication-store-worker';
+} from '@dineug/erd-editor-webview-bridge';
 import { encode } from 'base64-arraybuffer';
 
 const bridge = new Bridge();
