+++ packages/webview-bridge/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@dineug/erd-editor-vscode-bridge",
+  "name": "@dineug/erd-editor-webview-bridge",
   "version": "0.1.0",
   "private": true,
   "type": "module",
